# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/config/edit": {
      "post": {
        "tags": [
          "Config"
        ],
        "summary": "Modify the configuration about the torrent client",
        "description": "",
        "parameters": [
          {
            "name": "config",
            "in": "body",
            "description": "Configuration of the client",
            "schema": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string",
                  "example": "./"
                }
              },
              "required": [
                "path"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Set the path",
            "schema": {
              "type": "boolean",
              "example": true,
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
    "/stream/config": {
      "get": {
        "tags": [
          "Stream"
        ],
        "summary": "Retrieve the configuration about streaming",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Configuration data",
            "schema": {
              "type": "object",
              "properties": {
                "uptobox": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "xxxxxxx"
                    }
                  }
                }
              },
              "required": [
                "uptobox"
              ],
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Stream"
        ],
        "summary": "Update the configuration about streaming platform",
        "description": "",
        "parameters": [
          {
            "name": "conf",
            "in": "body",
            "description": "Data about configuration (Only uptobox available now",
            "schema": {
              "type": "object",
              "properties": {
                "uploadEnabled": {
                  "type": "boolean",
                  "example": false
                },
                "uptobox": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "xxxxxxx"
                    }
                  }
                }
              },
              "required": [
                "uploadEnabled",
                "uptobox"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Configuration data",
            "schema": {
              "type": "object",
              "properties": {
                "uptobox": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "xxxxxxx"
                    }
                  }
                }
              },
              "required": [
                "uptobox"
              ],
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/category/": {
      "get": {
        "tags": [
          "Category"
        ],
        "summary": "Get all the category that identify the default path for that category",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Category"
        ],
        "summary": "Create a category",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "example": "any"
                },
                "label": {
                  "example": "any"
                },
                "path": {
                  "example": "any"
                },
                "tag": {
                  "example": "any"
                },
                "tooltip": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "If the operation gone fine, it will return the id",
            "schema": {
              "type": "boolean",
              "example": true,
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Category"
        ],
        "summary": "Restore the default category",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "If the operation gone fine, it will return the id",
            "schema": {
              "type": "boolean",
              "example": true,
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/category/{categoryId}": {
      "put": {
        "tags": [
          "Category"
        ],
        "summary": "Edit a category, based on the id of the path",
        "description": "",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "example": "any"
                },
                "label": {
                  "example": "any"
                },
                "defaultSearch": {
                  "example": "any"
                },
                "path": {
                  "example": "any"
                },
                "tag": {
                  "example": "any"
                },
                "tooltip": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "If the operation gone fine, true",
            "schema": {
              "type": "boolean",
              "example": true,
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Category"
        ],
        "summary": "Remove a category, based on the id in the path param",
        "description": "",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "If the operation gone fine return true, otherwise false",
            "schema": {
              "type": "boolean",
              "example": true,
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}